+++ changelog.md
@@ -8,6 +8,7 @@ This version is not released yet and is under active development.
 
 - \[mpm\] Allow global `upgrade` of a subset of packages from the command line.
 - \[mpm\] Add a `-A/--all` option to `upgrade` operation.
+- \[mpm\] Alias `list` operation to `installed`, `uninstall` to `remove` and `update` to `upgrade`.
 - \[pipx\] Implement `outdated` operation.
 - \[mpm\] Rename all `*-like` labels to `*-based` to help finer identification of families.